+++ lib/game/df_camera.dart
@@ -26,4 +26,5 @@ class DFCamera {
     this.limit = limit;
     print("Camera setLimit:" + limit.toString());
   }
+
 }